feat: 「/」にアクセスすると滞在情報が出るようになった
