+++ src/components/layout/Layout.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { ReactNode } from 'react';
 import { useWindowSize } from 'usehooks-ts';
 import Foot from '@/components/layout/Footer';
