+++ src/utils/$path.ts
@@ -17,7 +17,9 @@ export const pagesPath = {
   userInformation: {
     $url: (url?: { hash?: string }) => ({ pathname: '/userInformation' as const, hash: url?.hash }),
   },
-  $url: (url?: { hash?: string }) => ({ pathname: '/' as const, hash: url?.hash }),
+  stayer: {
+    $url: (url?: { hash?: string }) => ({ pathname: '/stayer' as const, hash: url?.hash }),
+  }
 };
 
 export type PagesPath = typeof pagesPath;
